Fix working unresolved tree mode, fix processing cljc files corner...
...case

For the unresolved tree case flatten out the resolved tree and sort it
by the topological order of the fully expanded tree. Also use a
slightly different walking algorithm which unzips all dependencies
first, collects all the relevant contextual info and then processes
the source files in a reverse topological order.

Cljc file processing bug

In case a cljc files uses reader macro in its ns macro and refer to
the same namespace in clj and cljs which are apparently coming from
different source files we need to make sure that the inlined clj file
is only referred in the clj part of the reader macro while the inlined
cljs file is only referred in the cljs part.

cljfmt.v0v6v3.cljfmt.core (cljc) had this problem referring to both
rewrite-clj and rewrite-cljs