Bump MrAnderson deps
Two projects moved to clj-commons:
- me.raynes/fs -> clj-commons/fs
- com.cemerick/pomegranate -> clj-commons/pomegranate

Eastwood noticed that rewrite-clj has deprecated
`rewrite-clj.zip/edn`. This is a name change only to
`rewrite-clj.zip/of-node`.

Left Clojure version as is at 1.10.3 because it implies the minimum
Clojure version that MrAnderson supports.

The koacha cloverage plugin now checks for data readers on the
classpath. When running `:eval-in :leiningen`, this check fails with a
null pointer exception. Now doing an `:eval-in :sub-process` for
koacha profile.

Also added a `tests.edn` for koacha because it really wanted me to.

Closes #66